Fix compiler warning.
git-svn-id: http://svn.mythtv.org/svn/trunk@7378 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/dtvsignalmonitor.cpp
@@ -222,7 +222,7 @@ void DTVSignalMonitor::SetPMT(uint, const ProgramMapTable *pmt)
 {
     AddFlags(kDTVSigMon_PMTSeen);
 
-    if (pmt->ProgramNumber() != programNumber)
+    if (pmt->ProgramNumber() != (uint)programNumber)
         return; // We are not the PMT you are looking for.
 
     if (ignoreEncrypted && pmt->IsEncrypted())
